feat(billing): preserve termination metadata and price forgotten/removed STOREs
Support the credit service's v2 storage expense pipeline:

- forgotten_messages captures owner, payment_type (NULL coalesced to
  hold), file size and forgotten_at at forget time; forgotten_at is
  backfilled from still-indexed FORGET messages (migration 0061, indexes
  built concurrently, rerun-safe).
- New removed_messages table with a two-phase lifecycle: size snapshot
  when a message enters REMOVING (both balance jobs), row deleted on
  recovery, removed_at stamped by the garbage collector at
  REMOVING->REMOVED; all writes and denormalized status dual-writes are
  gated on the guarded status flip rowcount (migration 0063).
- messages.json supports terminal-status queries: msgStatuses=forgotten /
  removed window and sort on forgotten_at / removed_at, support
  paymentTypes (coalesced to hold for removed) and reject unsupported
  narrowing filters; single-message endpoints expose the new fields as
  epoch floats.
- /api/v0/price/{item_hash} computes live prices for forgotten and
  removed STOREs from the preserved metadata through the estimation path
  (same floor and pricing aggregate as live messages); missing metadata
  keeps returning 410 Gone.
- aleph_credit_expense v2 entries: optional count/size persisted to
  credit_history as expense_count / expense_size_mib (migration 0062).
- MIN_STORE_COST_MIB raised 25 -> 100 MiB (credit-only floor),
  documented including the transitional inconsistency with precomputed
  account_costs until a recalculation.

+++ src/aleph/schemas/cost_estimation_messages.py
@@ -86,7 +86,9 @@ class CostEstimationProgramContent(ProgramContent):
 
 
 class CostEstimationStoreContent(StoreContent):
-    estimated_size_mib: Optional[int] = None
+    # float so exact byte sizes (size / MiB) can be estimated, e.g. when
+    # pricing forgotten STORE messages from their preserved metadata.
+    estimated_size_mib: Optional[float] = None
 
 
 class BaseCostEstimationMessage(AlephBaseMessage, Generic[MType, ContentType]):
